refactor: improve code organization and modularity
Restructured the codebase for better maintainability and faster builds:

**New file structure:**
- Created `types/pomodoro.ts` for shared TypeScript types (Mode, PersistedPreferences, PomodoroSession)
- Created `lib/constants.ts` for app-wide constants (durations, storage keys)
- Created `lib/format.ts` for utility functions (formatTime)
- Created `lib/storage/preferences.ts` for preference persistence logic
- Created `lib/storage/sessions.ts` for session management logic

**Benefits:**
- Improved code reusability and DRY principles
- Better separation of concerns
- Easier testing and maintenance
- Faster build times through better tree-shaking
- Type safety across modules with shared types
- Cleaner imports in page.tsx (reduced from 293 to 209 lines)

**Changes:**
- Refactored `app/page.tsx` to use modular imports
- Updated `hooks/useTimer.ts` to use shared types
- Fixed deprecated `substr()` to `substring()` in session ID generation
- All functionality preserved, no breaking changes

Verified with successful build and typecheck ✓
