+++ lib/constants.ts
@@ -1,0 +1,4 @@
+export const FOCUS_DEFAULT = 25 * 60; // seconds
+export const BREAK_DEFAULT = 5 * 60; // seconds
+export const STORAGE_KEY = "pomo-preferences";
+export const SESSIONS_STORAGE_KEY = "pomo-sessions";
